+++ src/__tests__/api.test.ts
@@ -1,57 +1,18 @@
 import fs from 'fs';
 import { Api } from '../resources/Api';
-import { AppSyncConfig } from '../types/plugin';
-import Serverless from 'serverless/lib/Serverless';
-import { flatten, noop, set, upperFirst } from 'lodash';
-import { logger } from '../utils';
-import AwsProvider from 'serverless/lib/plugins/aws/provider.js';
-import ServerlessAppsyncPlugin from '..';
+import { flatten, upperFirst } from 'lodash';
 import globby from 'globby';
+import * as given from './given';
 
 // 2020-12-09T16:24:22+00:00
 jest.spyOn(Date, 'now').mockImplementation(() => 1607531062000);
 
-const config: AppSyncConfig = {
-  name: 'MyApi',
-  isSingleConfig: true,
-  xrayEnabled: false,
-  schema: ['schema.graphql'],
-  authentication: {
-    type: 'API_KEY',
-  },
-  additionalAuthenticationProviders: [],
-  mappingTemplatesLocation: {
-    resolvers: 'path/to/mappingTemplates',
-    pipelineFunctions: 'path/to/mappingTemplates',
-  },
-  resolvers: [],
-  pipelineFunctions: [],
-  dataSources: [],
-  substitutions: {},
-  tags: {
-    stage: 'Dev',
-  },
-};
-
-const serverless = new Serverless();
-serverless.setProvider('aws', new AwsProvider(serverless));
-serverless.serviceOutputs = new Map();
-serverless.servicePluginOutputs = new Map();
-set(serverless, 'configurationInput.custom.appSync', []);
-
-const options = {
-  stage: 'dev',
-  region: 'us-east-1',
-};
-const plugin = new ServerlessAppsyncPlugin(serverless, options, {
-  log: logger(noop),
-  writeText: noop,
-});
+const plugin = given.plugin();
 
 describe('Api', () => {
   describe('compileEndpoint', () => {
     it('should compile the Api Resoruce', () => {
-      const api = new Api(config, plugin);
+      const api = new Api(given.appSyncConfig(), plugin);
       expect(api.compileEndpoint()).toMatchInlineSnapshot(`
         Object {
           "GraphQlApi": Object {
@@ -74,14 +35,13 @@ describe('Api', () => {
 
     it('should compile the Api Resoruce with logs enabled', () => {
       const api = new Api(
-        {
-          ...config,
+        given.appSyncConfig({
           logConfig: {
             level: 'ERROR',
             excludeVerboseContent: false,
             logRetentionInDays: 14,
           },
-        },
+        }),
         plugin,
       );
       expect(api.compileEndpoint()).toMatchInlineSnapshot(`
@@ -116,8 +76,7 @@ describe('Api', () => {
 
     it('should compile the Api Resoruce with additional auths', () => {
       const api = new Api(
-        {
-          ...config,
+        given.appSyncConfig({
           additionalAuthenticationProviders: [
             {
               type: 'AMAZON_COGNITO_USER_POOLS',
@@ -148,7 +107,7 @@ describe('Api', () => {
               },
             },
           ],
-        },
+        }),
         plugin,
       );
       expect(api.compileEndpoint()).toMatchInlineSnapshot(`
@@ -210,8 +169,7 @@ describe('Api', () => {
 
   it('should use defaultAction as ALLOW for Cognito when primary auth and additionalAuth are present', () => {
     const api = new Api(
-      {
-        ...config,
+      given.appSyncConfig({
         authentication: {
           type: 'AMAZON_COGNITO_USER_POOLS',
           config: {
@@ -225,7 +183,7 @@ describe('Api', () => {
             type: 'AWS_IAM',
           },
         ],
-      },
+      }),
       plugin,
     );
     expect(api.compileEndpoint()).toMatchInlineSnapshot(`
@@ -261,8 +219,7 @@ describe('Api', () => {
 
   it('should use defaultAction as DENY for Cognito when primary auth and additionalAuth are not present', () => {
     const api = new Api(
-      {
-        ...config,
+      given.appSyncConfig({
         authentication: {
           type: 'AMAZON_COGNITO_USER_POOLS',
           config: {
@@ -272,7 +229,7 @@ describe('Api', () => {
           },
         },
         additionalAuthenticationProviders: [],
-      },
+      }),
       plugin,
     );
     expect(api.compileEndpoint()).toMatchInlineSnapshot(`
@@ -303,8 +260,7 @@ describe('Api', () => {
 
   it('should use defaultAction as DENY for Cognito when not primary auth', () => {
     const api = new Api(
-      {
-        ...config,
+      given.appSyncConfig({
         authentication: {
           type: 'API_KEY',
         },
@@ -318,7 +274,7 @@ describe('Api', () => {
             },
           },
         ],
-      },
+      }),
       plugin,
     );
     expect(api.compileEndpoint()).toMatchInlineSnapshot(`
@@ -360,9 +316,6 @@ describe('Api', () => {
         const matches = `${path}`.match(/([a-z]+)\.graphql$/);
         const type = upperFirst(matches?.[1] || 'Unknown');
         return `
-            """
-            ${path}
-            """
             type Query {
               get${type}: ${type}!
             }
@@ -393,7 +346,7 @@ describe('Api', () => {
 
     it('should merge the schemas', () => {
       const api = new Api(
-        { ...config, schema: ['users.graphql', 'posts.graphql'] },
+        given.appSyncConfig({ schema: ['users.graphql', 'posts.graphql'] }),
         plugin,
       );
       expect(api.compileSchema()).toMatchInlineSnapshot(`
@@ -406,8 +359,7 @@ describe('Api', () => {
                   "ApiId",
                 ],
               },
-              "Definition": "# posts.graphql
-        type Query  {
+              "Definition": "type Query  {
           getUsers: Users!
           getPosts: Posts!
         }
@@ -427,7 +379,10 @@ describe('Api', () => {
     });
 
     it('should merge glob schemas', () => {
-      const api = new Api({ ...config, schema: ['*.graphql'] }, plugin);
+      const api = new Api(
+        given.appSyncConfig({ schema: ['*.graphql'] }),
+        plugin,
+      );
       expect(api.compileSchema()).toMatchInlineSnapshot(`
         Object {
           "GraphQlSchema": Object {
@@ -438,8 +393,7 @@ describe('Api', () => {
                   "ApiId",
                 ],
               },
-              "Definition": "# posts.graphql
-        type Query  {
+              "Definition": "type Query  {
           getUsers: Users!
           getPosts: Posts!
         }
@@ -461,7 +415,7 @@ describe('Api', () => {
 
   describe('CloudWatch', () => {
     it('should not compile CloudWatch Resources when disabled', () => {
-      const api = new Api(config, plugin);
+      const api = new Api(given.appSyncConfig(), plugin);
       expect(api.compileCloudWatchLogGroup()).toMatchInlineSnapshot(
         `Object {}`,
       );
@@ -469,13 +423,12 @@ describe('Api', () => {
 
     it('should compile CloudWatch Resources when enaabled', () => {
       const api = new Api(
-        {
-          ...config,
+        given.appSyncConfig({
           logConfig: {
             level: 'ERROR',
             logRetentionInDays: 14,
           },
-        },
+        }),
         plugin,
       );
       expect(api.compileCloudWatchLogGroup()).toMatchInlineSnapshot(`
@@ -528,8 +481,7 @@ describe('Api', () => {
 
   describe('apiKeys', () => {
     const api = new Api(
-      {
-        ...config,
+      given.appSyncConfig({
         apiKeys: [
           {
             name: 'Default',
@@ -557,7 +509,7 @@ describe('Api', () => {
           },
           'InlineKey',
         ],
-      },
+      }),
       plugin,
     );
     it('should generate api keys', () => {
@@ -679,12 +631,11 @@ describe('Api', () => {
   describe('LambdaAuthorizer', () => {
     it('should not generate the Lambda Authorizer Resoruces', () => {
       const api = new Api(
-        {
-          ...config,
+        given.appSyncConfig({
           authentication: {
             type: 'API_KEY',
           },
-        },
+        }),
         plugin,
       );
       expect(api.compileLambdaAuthorizerPermission()).toMatchInlineSnapshot(
@@ -694,15 +645,14 @@ describe('Api', () => {
 
     it('should generate the Lambda Authorizer Resoruces from basic auth', () => {
       const api = new Api(
-        {
-          ...config,
+        given.appSyncConfig({
           authentication: {
             type: 'AWS_LAMBDA',
             config: {
               functionArn: 'arn:',
             },
           },
-        },
+        }),
         plugin,
       );
       expect(api.compileLambdaAuthorizerPermission()).toMatchInlineSnapshot(`
@@ -724,8 +674,7 @@ describe('Api', () => {
 
     it('should generate the Lambda Authorizer Resources from additional auth', () => {
       const api = new Api(
-        {
-          ...config,
+        given.appSyncConfig({
           additionalAuthenticationProviders: [
             {
               type: 'AWS_LAMBDA',
@@ -734,7 +683,7 @@ describe('Api', () => {
               },
             },
           ],
-        },
+        }),
         plugin,
       );
       expect(api.compileLambdaAuthorizerPermission()).toMatchInlineSnapshot(`
@@ -754,467 +703,21 @@ describe('Api', () => {
       `);
     });
   });
-
-  describe('Resolvers', () => {
-    let mock: jest.SpyInstance;
-    beforeAll(() => {
-      mock = jest
-        .spyOn(fs, 'readFileSync')
-        .mockImplementation(
-          (path) => `Content of ${`${path}`.replace(/\\/g, '/')}`,
-        );
-    });
-
-    afterAll(() => {
-      mock.mockRestore();
-    });
-
-    describe('Unit Resolvers', () => {
-      it('should generate Resources with default mapping templates', () => {
-        const api = new Api(config, plugin);
-        expect(
-          api.compileResolver({
-            dataSource: 'myTable',
-            kind: 'UNIT',
-            type: 'Query',
-            field: 'user',
-          }),
-        ).toMatchInlineSnapshot(`
-          Object {
-            "GraphQlResolverQueryuser": Object {
-              "DependsOn": Array [
-                "GraphQlSchema",
-              ],
-              "Properties": Object {
-                "ApiId": Object {
-                  "Fn::GetAtt": Array [
-                    "GraphQlApi",
-                    "ApiId",
-                  ],
-                },
-                "DataSourceName": "myTable",
-                "FieldName": "user",
-                "Kind": "UNIT",
-                "RequestMappingTemplate": "Content of path/to/mappingTemplates/Query.user.request.vtl",
-                "ResponseMappingTemplate": "Content of path/to/mappingTemplates/Query.user.response.vtl",
-                "TypeName": "Query",
-              },
-              "Type": "AWS::AppSync::Resolver",
-            },
-          }
-        `);
-      });
-
-      it('should generate Resources with default specific templates', () => {
-        const api = new Api(config, plugin);
-        expect(
-          api.compileResolver({
-            dataSource: 'myTable',
-            kind: 'UNIT',
-            type: 'Query',
-            field: 'user',
-            request: 'specific.request.tpl',
-            response: 'specific.response.tpl',
-          }),
-        ).toMatchInlineSnapshot(`
-          Object {
-            "GraphQlResolverQueryuser": Object {
-              "DependsOn": Array [
-                "GraphQlSchema",
-              ],
-              "Properties": Object {
-                "ApiId": Object {
-                  "Fn::GetAtt": Array [
-                    "GraphQlApi",
-                    "ApiId",
-                  ],
-                },
-                "DataSourceName": "myTable",
-                "FieldName": "user",
-                "Kind": "UNIT",
-                "RequestMappingTemplate": "Content of path/to/mappingTemplates/specific.request.tpl",
-                "ResponseMappingTemplate": "Content of path/to/mappingTemplates/specific.response.tpl",
-                "TypeName": "Query",
-              },
-              "Type": "AWS::AppSync::Resolver",
-            },
-          }
-        `);
-      });
-
-      it('should generate Resources with direct Lambda templates', () => {
-        const api = new Api(config, plugin);
-        expect(
-          api.compileResolver({
-            dataSource: 'myLambdaFunction',
-            kind: 'UNIT',
-            type: 'Query',
-            field: 'user',
-            request: 'specific.request.tpl',
-            response: 'specific.response.tpl',
-          }),
-        ).toMatchInlineSnapshot(`
-          Object {
-            "GraphQlResolverQueryuser": Object {
-              "DependsOn": Array [
-                "GraphQlSchema",
-              ],
-              "Properties": Object {
-                "ApiId": Object {
-                  "Fn::GetAtt": Array [
-                    "GraphQlApi",
-                    "ApiId",
-                  ],
-                },
-                "DataSourceName": "myLambdaFunction",
-                "FieldName": "user",
-                "Kind": "UNIT",
-                "RequestMappingTemplate": "Content of path/to/mappingTemplates/specific.request.tpl",
-                "ResponseMappingTemplate": "Content of path/to/mappingTemplates/specific.response.tpl",
-                "TypeName": "Query",
-              },
-              "Type": "AWS::AppSync::Resolver",
-            },
-          }
-        `);
-      });
-    });
-
-    describe('Pipeline Resovlers', () => {
-      it('should generate Resources with default mapping templates', () => {
-        const api = new Api(config, plugin);
-        expect(
-          api.compileResolver({
-            kind: 'PIPELINE',
-            type: 'Query',
-            field: 'user',
-            functions: ['function1', 'function2'],
-          }),
-        ).toMatchInlineSnapshot(`
-          Object {
-            "GraphQlResolverQueryuser": Object {
-              "DependsOn": Array [
-                "GraphQlSchema",
-              ],
-              "Properties": Object {
-                "ApiId": Object {
-                  "Fn::GetAtt": Array [
-                    "GraphQlApi",
-                    "ApiId",
-                  ],
-                },
-                "FieldName": "user",
-                "Kind": "PIPELINE",
-                "PipelineConfig": Object {
-                  "Functions": Array [
-                    Object {
-                      "Fn::GetAtt": Array [
-                        "GraphQlFunctionConfigurationfunction1",
-                        "FunctionId",
-                      ],
-                    },
-                    Object {
-                      "Fn::GetAtt": Array [
-                        "GraphQlFunctionConfigurationfunction2",
-                        "FunctionId",
-                      ],
-                    },
-                  ],
-                },
-                "RequestMappingTemplate": "Content of path/to/mappingTemplates/Query.user.request.vtl",
-                "ResponseMappingTemplate": "Content of path/to/mappingTemplates/Query.user.response.vtl",
-                "TypeName": "Query",
-              },
-              "Type": "AWS::AppSync::Resolver",
-            },
-          }
-        `);
-      });
-
-      it('should generate Resources with specific mapping templates', () => {
-        const api = new Api(config, plugin);
-        expect(
-          api.compileResolver({
-            kind: 'PIPELINE',
-            type: 'Query',
-            field: 'user',
-            functions: ['function1', 'function2'],
-            request: 'specific.request.tpl',
-            response: 'specific.response.tpl',
-          }),
-        ).toMatchInlineSnapshot(`
-          Object {
-            "GraphQlResolverQueryuser": Object {
-              "DependsOn": Array [
-                "GraphQlSchema",
-              ],
-              "Properties": Object {
-                "ApiId": Object {
-                  "Fn::GetAtt": Array [
-                    "GraphQlApi",
-                    "ApiId",
-                  ],
-                },
-                "FieldName": "user",
-                "Kind": "PIPELINE",
-                "PipelineConfig": Object {
-                  "Functions": Array [
-                    Object {
-                      "Fn::GetAtt": Array [
-                        "GraphQlFunctionConfigurationfunction1",
-                        "FunctionId",
-                      ],
-                    },
-                    Object {
-                      "Fn::GetAtt": Array [
-                        "GraphQlFunctionConfigurationfunction2",
-                        "FunctionId",
-                      ],
-                    },
-                  ],
-                },
-                "RequestMappingTemplate": "Content of path/to/mappingTemplates/specific.request.tpl",
-                "ResponseMappingTemplate": "Content of path/to/mappingTemplates/specific.response.tpl",
-                "TypeName": "Query",
-              },
-              "Type": "AWS::AppSync::Resolver",
-            },
-          }
-        `);
-      });
-    });
-
-    describe('Pipeline Function', () => {
-      it('should generate Pipeline Function Resources with default mapping templates', () => {
-        const api = new Api(config, plugin);
-        expect(
-          api.compilePipelineFunctionResource({
-            name: 'function1',
-            dataSource: 'myTable',
-            description: 'Function1 Pipeline Resolver',
-          }),
-        ).toMatchInlineSnapshot(`
-          Object {
-            "GraphQlFunctionConfigurationfunction1": Object {
-              "Properties": Object {
-                "ApiId": Object {
-                  "Fn::GetAtt": Array [
-                    "GraphQlApi",
-                    "ApiId",
-                  ],
-                },
-                "DataSourceName": Object {
-                  "Fn::GetAtt": Array [
-                    "GraphQlDsmyTable",
-                    "Name",
-                  ],
-                },
-                "Description": "Function1 Pipeline Resolver",
-                "FunctionVersion": "2018-05-29",
-                "Name": "function1",
-                "RequestMappingTemplate": "Content of path/to/mappingTemplates/function1.request.vtl",
-                "ResponseMappingTemplate": "Content of path/to/mappingTemplates/function1.response.vtl",
-              },
-              "Type": "AWS::AppSync::FunctionConfiguration",
-            },
-          }
-        `);
-      });
-
-      it('should generate Pipeline Function Resources with specific mapping tempaltes', () => {
-        const api = new Api(config, plugin);
-        expect(
-          api.compilePipelineFunctionResource({
-            name: 'function1',
-            dataSource: 'myTable',
-            description: 'Function1 Pipeline Resolver',
-            request: 'specific.request.tpl',
-            response: 'specific.response.tpl',
-          }),
-        ).toMatchInlineSnapshot(`
-          Object {
-            "GraphQlFunctionConfigurationfunction1": Object {
-              "Properties": Object {
-                "ApiId": Object {
-                  "Fn::GetAtt": Array [
-                    "GraphQlApi",
-                    "ApiId",
-                  ],
-                },
-                "DataSourceName": Object {
-                  "Fn::GetAtt": Array [
-                    "GraphQlDsmyTable",
-                    "Name",
-                  ],
-                },
-                "Description": "Function1 Pipeline Resolver",
-                "FunctionVersion": "2018-05-29",
-                "Name": "function1",
-                "RequestMappingTemplate": "Content of path/to/mappingTemplates/specific.request.tpl",
-                "ResponseMappingTemplate": "Content of path/to/mappingTemplates/specific.response.tpl",
-              },
-              "Type": "AWS::AppSync::FunctionConfiguration",
-            },
-          }
-        `);
-      });
-
-      it('should generate Pipeline Function Resources with direct Lambda mapping tempaltes', () => {
-        const api = new Api(config, plugin);
-        expect(
-          api.compilePipelineFunctionResource({
-            name: 'function1',
-            dataSource: 'myLambdaFunction',
-            description: 'Function1 Pipeline Resolver',
-            request: false,
-            response: false,
-          }),
-        ).toMatchInlineSnapshot(`
-          Object {
-            "GraphQlFunctionConfigurationfunction1": Object {
-              "Properties": Object {
-                "ApiId": Object {
-                  "Fn::GetAtt": Array [
-                    "GraphQlApi",
-                    "ApiId",
-                  ],
-                },
-                "DataSourceName": Object {
-                  "Fn::GetAtt": Array [
-                    "GraphQlDsmyLambdaFunction",
-                    "Name",
-                  ],
-                },
-                "Description": "Function1 Pipeline Resolver",
-                "FunctionVersion": "2018-05-29",
-                "Name": "function1",
-              },
-              "Type": "AWS::AppSync::FunctionConfiguration",
-            },
-          }
-        `);
-      });
-    });
-
-    describe('Caching', () => {
-      it('should generate Resources with caching enabled', () => {
-        const api = new Api(
-          {
-            ...config,
-            caching: {
-              behavior: 'PER_RESOLVER_CACHING',
-            },
-          },
-          plugin,
-        );
-        expect(
-          api.compileResolver({
-            dataSource: 'myTable',
-            kind: 'UNIT',
-            type: 'Query',
-            field: 'user',
-            caching: true,
-          }),
-        ).toMatchInlineSnapshot(`
-          Object {
-            "GraphQlResolverQueryuser": Object {
-              "DependsOn": Array [
-                "GraphQlSchema",
-              ],
-              "Properties": Object {
-                "ApiId": Object {
-                  "Fn::GetAtt": Array [
-                    "GraphQlApi",
-                    "ApiId",
-                  ],
-                },
-                "CachingConfig": Object {
-                  "Ttl": 3600,
-                },
-                "DataSourceName": "myTable",
-                "FieldName": "user",
-                "Kind": "UNIT",
-                "RequestMappingTemplate": "Content of path/to/mappingTemplates/Query.user.request.vtl",
-                "ResponseMappingTemplate": "Content of path/to/mappingTemplates/Query.user.response.vtl",
-                "TypeName": "Query",
-              },
-              "Type": "AWS::AppSync::Resolver",
-            },
-          }
-        `);
-      });
-
-      it('should generate Resources with custom keys', () => {
-        const api = new Api(
-          {
-            ...config,
-            caching: {
-              behavior: 'PER_RESOLVER_CACHING',
-            },
-          },
-          plugin,
-        );
-        expect(
-          api.compileResolver({
-            dataSource: 'myTable',
-            kind: 'UNIT',
-            type: 'Query',
-            field: 'user',
-            caching: {
-              ttl: 200,
-              keys: ['$context.identity.sub', '$context.arguments.id'],
-            },
-          }),
-        ).toMatchInlineSnapshot(`
-          Object {
-            "GraphQlResolverQueryuser": Object {
-              "DependsOn": Array [
-                "GraphQlSchema",
-              ],
-              "Properties": Object {
-                "ApiId": Object {
-                  "Fn::GetAtt": Array [
-                    "GraphQlApi",
-                    "ApiId",
-                  ],
-                },
-                "CachingConfig": Object {
-                  "CachingKeys": Array [
-                    "$context.identity.sub",
-                    "$context.arguments.id",
-                  ],
-                  "Ttl": 200,
-                },
-                "DataSourceName": "myTable",
-                "FieldName": "user",
-                "Kind": "UNIT",
-                "RequestMappingTemplate": "Content of path/to/mappingTemplates/Query.user.request.vtl",
-                "ResponseMappingTemplate": "Content of path/to/mappingTemplates/Query.user.response.vtl",
-                "TypeName": "Query",
-              },
-              "Type": "AWS::AppSync::Resolver",
-            },
-          }
-        `);
-      });
-    });
-  });
 });
 
 describe('Caching', () => {
   it('should not generate Resources when disabled', () => {
-    const api = new Api({ ...config, caching: undefined }, plugin);
+    const api = new Api(given.appSyncConfig({ caching: undefined }), plugin);
     expect(api.compileCachingResources()).toEqual({});
   });
 
   it('should generate Resources with defaults', () => {
     const api = new Api(
-      {
-        ...config,
+      given.appSyncConfig({
         caching: {
           behavior: 'FULL_REQUEST_CACHING',
         },
-      },
+      }),
       plugin,
     );
     expect(api.compileCachingResources()).toMatchInlineSnapshot(`
@@ -1241,8 +744,7 @@ describe('Caching', () => {
 
   it('should generate Resources with custom Config', () => {
     const api = new Api(
-      {
-        ...config,
+      given.appSyncConfig({
         caching: {
           behavior: 'FULL_REQUEST_CACHING',
           atRestEncryption: true,
@@ -1250,7 +752,7 @@ describe('Caching', () => {
           ttl: 500,
           type: 'T2_MEDIUM',
         },
-      },
+      }),
       plugin,
     );
     expect(api.compileCachingResources()).toMatchInlineSnapshot(`
